Remove tool source duplication Various nits across codebase
