+++ src/interfaces/hcorenum.hpp
@@ -87,7 +87,7 @@ private:
     HRESULT ResetDynamicEnum(_In_ uint32_t position) noexcept;
 };
 
-struct HCORENUMImplDeleter
+struct HCORENUMImplDeleter final
 {
     using pointer = HCORENUMImpl*;
     void operator()(HCORENUMImpl* mem)
